split up the record game modal
