+++ board-game-statistics/src/pages/group/games/steps/PlayersStep.tsx
@@ -1,0 +1,219 @@
+import Form from 'react-bootstrap/Form';
+
+type WinCondition = 'single' | 'team';
+
+interface Player {
+	id: string;
+	name: string;
+}
+
+interface PlayersStepProps {
+	groupPlayers: Player[];
+	selectedPlayerIds: string[];
+	onTogglePlayer: (playerId: string) => void;
+	winCondition: WinCondition;
+	numTeams: number | '';
+	playerIdToTeam: Record<string, string>;
+	onPlayerTeamChange: (playerId: string, team: string) => void;
+	playerPoints: Record<string, number>;
+	onPlayerPointsChange: (playerId: string, points: number) => void;
+	singleWinnerId: string;
+	onSingleWinnerChange: (playerId: string) => void;
+	teamWinner: string;
+	onTeamWinnerChange: (team: string) => void;
+}
+
+const PlayersStep = (props: PlayersStepProps) => {
+	const {
+		groupPlayers,
+		selectedPlayerIds,
+		onTogglePlayer,
+		winCondition,
+		numTeams,
+		playerIdToTeam,
+		onPlayerTeamChange,
+		playerPoints,
+		onPlayerPointsChange,
+		singleWinnerId,
+		onSingleWinnerChange,
+		teamWinner,
+		onTeamWinnerChange,
+	} = props;
+
+	return (
+		<div className='vstack gap-2'>
+			<Form>
+				<Form.Group>
+					<Form.Label>Select players</Form.Label>
+					<div className='d-flex flex-column'>
+						{groupPlayers.map((p) => (
+							<div
+								key={p.id}
+								className='d-flex align-items-center gap-2 mb-1'
+							>
+								<Form.Check
+									type='checkbox'
+									label={p.name}
+									checked={selectedPlayerIds.includes(p.id)}
+									onChange={() => onTogglePlayer(p.id)}
+								/>
+								{winCondition === 'team' &&
+									selectedPlayerIds.includes(p.id) && (
+										<Form.Select
+											aria-label='Select team'
+											size='sm'
+											className='w-auto'
+											value={playerIdToTeam[p.id] ?? '1'}
+											onChange={(e) =>
+												onPlayerTeamChange(
+													p.id,
+													e.target.value
+												)
+											}
+										>
+											{(numTeams || 0) >= 2
+												? Array.from(
+														{
+															length: Number(
+																numTeams
+															),
+														},
+														(_, i) => `${i + 1}`
+													).map((t) => (
+														<option
+															key={t}
+															value={t}
+														>
+															{`Team ${t}`}
+														</option>
+													))
+												: [
+														<option
+															key='1'
+															value='1'
+														>
+															Team 1
+														</option>,
+														<option
+															key='2'
+															value='2'
+														>
+															Team 2
+														</option>,
+													]}
+										</Form.Select>
+									)}
+								{selectedPlayerIds.includes(p.id) && (
+									<Form.Control
+										type='number'
+										size='sm'
+										className='w-auto ms-2'
+										placeholder='Points'
+										value={playerPoints[p.id] || ''}
+										onChange={(e) =>
+											onPlayerPointsChange(
+												p.id,
+												Number(e.target.value) || 0
+											)
+										}
+									/>
+								)}
+							</div>
+						))}
+					</div>
+				</Form.Group>
+				{winCondition === 'team' &&
+					(numTeams === '' || Number(numTeams) < 2) && (
+						<div className='text-danger small mt-1'>
+							Enter number of teams (2 or more).
+						</div>
+					)}
+				{winCondition === 'team' &&
+					Number(numTeams) >= 2 &&
+					selectedPlayerIds.length > 0 &&
+					selectedPlayerIds.some((pid) => !playerIdToTeam[pid]) && (
+						<div className='text-danger small mt-1'>
+							Assign a team to all selected players.
+						</div>
+					)}
+				{winCondition === 'team' &&
+					Number(numTeams) >= 2 &&
+					(() => {
+						const totals = new Array(Number(numTeams)).fill(0);
+						selectedPlayerIds.forEach((pid) => {
+							const t = Number(playerIdToTeam[pid] || '0');
+							if (t >= 1 && t <= Number(numTeams)) {
+								totals[t - 1]++;
+							}
+						});
+						return totals.some((c) => c === 0);
+					})() && (
+						<div className='text-danger small mt-1'>
+							Each team must have at least one player.
+						</div>
+					)}
+				{winCondition === 'single' && selectedPlayerIds.length > 0 && (
+					<Form.Group className='mt-2'>
+						<Form.Label>Winning player</Form.Label>
+						<div>
+							{selectedPlayerIds.map((pid) => (
+								<Form.Check
+									key={`singleWinner-${pid}`}
+									inline
+									type='radio'
+									label={
+										groupPlayers.find((p) => p.id === pid)
+											?.name || pid
+									}
+									name='singleWinner'
+									id={`singleWinner-${pid}`}
+									checked={singleWinnerId === pid}
+									onChange={() => onSingleWinnerChange(pid)}
+								/>
+							))}
+						</div>
+					</Form.Group>
+				)}
+				{winCondition === 'team' && selectedPlayerIds.length > 0 && (
+					<Form.Group className='mt-2'>
+						<Form.Label>Winning team</Form.Label>
+						<div>
+							{(numTeams || 0) >= 2 &&
+								Array.from(
+									{ length: Number(numTeams) },
+									(_, i) => `${i + 1}`
+								).map((t) => (
+									<Form.Check
+										key={`teamWinner-${t}`}
+										inline
+										type='radio'
+										label={`Team ${t}`}
+										name='teamWinner'
+										id={`teamWinner-${t}`}
+										checked={teamWinner === t}
+										onChange={() => onTeamWinnerChange(t)}
+									/>
+								))}
+						</div>
+					</Form.Group>
+				)}
+				{winCondition === 'single' &&
+					selectedPlayerIds.length > 0 &&
+					singleWinnerId === '' && (
+						<div className='text-danger small mt-1'>
+							Select a winning player.
+						</div>
+					)}
+				{winCondition === 'team' &&
+					selectedPlayerIds.length > 0 &&
+					teamWinner === '' && (
+						<div className='text-danger small mt-1'>
+							Select a winning team.
+						</div>
+					)}
+			</Form>
+		</div>
+	);
+};
+
+export default PlayersStep;
